Fixed compilation warning on BSD.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@55640 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/test/test_ncbifile.cpp
@@ -1422,7 +1422,7 @@ static void s_TEST_FileIO_LargeFiles(void)
     assert(getrlimit(RLIMIT_FSIZE, &rl) == 0);
     cout << "File size limits = (" << rl.rlim_cur << ", " 
                                    << rl.rlim_max << ")" << endl;
-    assert(rl.rlim_cur > kSize_6GB);
+    assert((Uint8)rl.rlim_cur > kSize_6GB);
 #endif
     cout << endl;
